+++ apps/explorer/src/hooks/useGetOrders.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 
-import { ALL_SUPPORTED_CHAIN_IDS } from '@cowprotocol/cow-sdk'
+import { ALL_SUPPORTED_CHAIN_IDS, getAddressKey } from '@cowprotocol/cow-sdk'
 
 import { Props as ExplorerLinkProps } from 'components/common/BlockExplorerLink'
 import { useMultipleErc20 } from 'hooks/useErc20'
@@ -95,8 +95,8 @@ function useOrdersWithTokenInfo(networkId: Network | undefined): UseOrdersWithTo
     }
 
     const newOrders = orders.map((order) => {
-      order.buyToken = valueErc20s[order.buyTokenAddress.toLowerCase()] || order.buyToken
-      order.sellToken = valueErc20s[order.sellTokenAddress.toLowerCase()] || order.sellToken
+      order.buyToken = valueErc20s[getAddressKey(order.buyTokenAddress)] || order.buyToken
+      order.sellToken = valueErc20s[getAddressKey(order.sellTokenAddress)] || order.sellToken
 
       return order
     })