+++ apps/widget-configurator/src/app/configurator/types.ts
@@ -1,5 +1,11 @@
 import type { SupportedChainId } from '@cowprotocol/cow-sdk'
-import { CowSwapWidgetPaletteColors, PartnerFee, SlippageConfig, TradeType } from '@cowprotocol/widget-lib'
+import {
+  CowSwapWidgetPaletteColors,
+  PartnerFee,
+  SlippageConfig,
+  TradeType,
+  WidgetHookEvents,
+} from '@cowprotocol/widget-lib'
 
 import { PaletteMode } from '@mui/material'
 
@@ -10,13 +16,18 @@ export type ColorPalette = {
 export interface TokenListItem {
   url: string
   enabled: boolean
+  enabledForSell: boolean
+  enabledForBuy: boolean
 }
 
 export interface ConfiguratorState {
   chainId?: SupportedChainId
+  locale?: string
   theme: PaletteMode
+  boxShadow?: string
   currentTradeType: TradeType
   enabledTradeTypes: TradeType[]
+  enabledWidgetHooks: WidgetHookEvents[]
   sellToken: string
   sellTokenAmount: number | undefined
   buyToken: string
@@ -33,7 +44,14 @@ export interface ConfiguratorState {
   standaloneMode: boolean
   disableToastMessages: boolean
   disableProgressBar: boolean
+  disablePostTradeTips: boolean
+  disableCrossChainSwap: boolean
+  disableTokenImport: boolean
+  hideRecentTokens: boolean
+  hideFavoriteTokens: boolean
   hideBridgeInfo: boolean | undefined
   hideOrdersTable: boolean | undefined
+  disableTradeWhenPriceImpactIsUnknown: boolean
+  disableTradeWhenPriceImpactIsHigherThan: number | undefined
   slippage?: SlippageConfig
 }